requirements.txt
ultralytics==8.0.145
filetype==1.2.0
Flask==2.3.2
Pillow==9.5.0
easyocr==1.7.0